+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## [Unrelease]
 ### [Added]
 - Add the head file `includes/LoggingCpp.h` to define the namespace and class.
+- Define the method `LoggingCpp::myLogger::getCurrentTimestamp()` to get the current timestamp.
 
 ### [Changed]